+++ test.js
@@ -10,13 +10,15 @@ const {
     dictKeys,
     isObject,
     Signature,
+    HMACSHA256Signature,
+    HMACSHA512Signature,
     extractSignedData,
     makeValidUntil,
     validateSignature,
     DEFAULT_SIGNATURE_PARAM,
     validateSignedRequestData,
+    defaultValueDumper,
 } = require("./commonjs/index.js");
-// import {getSignatureData} from "./examples";
 
 /**
  * Helper functions for testing.
@@ -487,6 +489,185 @@ test("Test generateSignature", (t) => {
         }
     );
     t.deepEqual(signature3, expectedSignature3);
+
+    // Integration tests
+    // Signature test case 3a
+    const signature3a = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": "â"}
+    );
+    const expectedSignature3a = new Signature(
+        "dlT2WO/jYq7+xcvDEUkCnNW5TxA=",
+        "me@example.com",
+        "1628717009.0",
+        {"one": "â"}
+    );
+    t.deepEqual(signature3a, expectedSignature3a);
+
+    // Signature test case 4
+    const signature4 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": {"value": "â"}}
+    );
+    const expectedSignature4 = new Signature(
+        "+pA63D4EMF2pcfIlE/dYXyNkhx4=",
+        "me@example.com",
+        "1628717009.0",
+        {"one": {"value": "â"}}
+    );
+    t.deepEqual(signature4, expectedSignature4);
+
+    // SHA256 signatures
+    // Signature test case 11
+    const signature11 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const expectedSignature11 = new Signature(
+        "EZ7uXeeopIxK3wL62J/9tKPXoGmNk9V3KHGgwge9/ek=",
+        "me@example.com",
+        "1628717009.0",
+        {}
+    );
+    t.deepEqual(signature11, expectedSignature11);
+
+    // Signature test case 12
+    const signature12 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": "1", "two": "2"},
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const expectedSignature12 = new Signature(
+        "Cl90LfQ2L3DW2MAhZriqCfEisPdL+1aHA/M0GPc1Yr4=",
+        "me@example.com",
+        "1628717009.0",
+        {"one": "1", "two": "2"}
+    );
+    t.deepEqual(signature12, expectedSignature12);
+
+    // Signature test case 13
+    const signature13 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": "â"},
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const expectedSignature13 = new Signature(
+        "9UpLTlFgEbCJ2C4/gC4eDogn0JiuMzo7osbMEOejwkQ=",
+        "me@example.com",
+        "1628717009.0",
+        {"one": "â"}
+    );
+    t.deepEqual(signature13, expectedSignature13);
+
+    // Signature test case 14
+    const signature14 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": {"value": "â"}},
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const expectedSignature14 = new Signature(
+        "9Tg3PdJYm/2tKZtVU0F/5T6TtL39Rwy4Uniq36ZClMY=",
+        "me@example.com",
+        "1628717009.0",
+        {"one": {"value": "â"}}
+    );
+    t.deepEqual(signature14, expectedSignature14);
+
+    // SHA512 signatures
+    // Signature test case 21
+    const signature21 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const expectedSignature21 = new Signature(
+        "7QcInLFxLrv1TeZZY4EXbAc1YguBlcjmYfFe5J+FH4TAOquSBZvKwYLSQCS4VVmdhDDU1h1zVlPDc4MAW6SHGQ==",
+        "me@example.com",
+        "1628717009.0",
+        {}
+    );
+    t.deepEqual(signature21, expectedSignature21);
+
+    // Signature test case 22
+    const signature22 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": "1", "two": "2"},
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const expectedSignature22 = new Signature(
+        "+Bm5xtd3Cl+7VV0RM6H14z68M8vWuMP168m3UsXLP1jHTTQCg3mXxTncZ9a57AoQefh/qNmDdnD5AmFYGzJ+PQ==",
+        "me@example.com",
+        "1628717009.0",
+        {"one": "1", "two": "2"}
+    );
+    t.deepEqual(signature22, expectedSignature22);
+
+    // Signature test case 23
+    const signature23 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": "â"},
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const expectedSignature23 = new Signature(
+        "yockrWxDncGJ2/HMEi/ma/auEmv8xlIMm5U50CuTFYSKbzrgNPh4OXgax/s2d96+paaLagwmnZK1+xUGHeArXw==",
+        "me@example.com",
+        "1628717009.0",
+        {"one": "â"}
+    );
+    t.deepEqual(signature23, expectedSignature23);
+
+    // Signature test case 14
+    const signature24 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        {"one": {"value": "â"}},
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const expectedSignature24 = new Signature(
+        "OlFZzu/SlBQYWny3CVvP7ghiL6X8G4r/yS9yNl8N+9b1arae3AkMLCp+0MuLs2sp8qdM3j+a7MYdCQCBSOnAoQ==",
+        "me@example.com",
+        "1628717009.0",
+        {"one": {"value": "â"}}
+    );
+    t.deepEqual(signature24, expectedSignature24);
 });
 
 test("Test signatureToDict", (t) => {
@@ -889,6 +1070,96 @@ test("Test validateSignature", (t) => {
     );
     t.is(isValidSignature2, false);
     t.is(signature2.isExpired(), true);
+
+    // Test case 21 - valid non-expired signature
+    const signature21 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        makeValidUntil(),
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const isValidSignature21 = validateSignature(
+        signature21.signature,
+        signature21.authUser,
+        SECRET_KEY,
+        signature21.validUntil,
+        signature21.extra,
+        false,
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    t.is(isValidSignature21, true);
+
+    // Test case 22 - expired signature
+    const signature22 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    const isValidSignature22 = validateSignature(
+        signature22.signature,
+        signature22.authUser,
+        SECRET_KEY,
+        signature22.validUntil,
+        signature22.extra,
+        false,
+        defaultValueDumper,
+        HMACSHA256Signature,
+    );
+    t.is(isValidSignature22, false);
+    t.is(signature22.isExpired(), true);
+
+    // Test case 31 - valid non-expired signature
+    const signature31 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        makeValidUntil(),
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const isValidSignature31 = validateSignature(
+        signature31.signature,
+        signature31.authUser,
+        SECRET_KEY,
+        signature31.validUntil,
+        signature31.extra,
+        false,
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    t.is(isValidSignature31, true);
+
+    // Test case 32 - expired signature
+    const signature32 = generateSignature(
+        AUTH_USER,
+        SECRET_KEY,
+        validUntil,
+        SIGNATURE_LIFETIME,
+        null,
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    const isValidSignature32 = validateSignature(
+        signature32.signature,
+        signature32.authUser,
+        SECRET_KEY,
+        signature32.validUntil,
+        signature32.extra,
+        false,
+        defaultValueDumper,
+        HMACSHA512Signature,
+    );
+    t.is(isValidSignature32, false);
+    t.is(signature32.isExpired(), true);
 });
 
 test("Test validateSignedRequestData", (t) => {